requirements.txt
Pillow==9.2.0
rel==0.4.7
requests==2.28.1
websocket-client==1.4.1
